+++ test/user-query-test.ts
@@ -4,7 +4,6 @@ import { prisma } from "../src/setup-db";
 import { hashPassword } from "../src/graphql/resolvers";
 import { generateToken } from "../src/graphql/helpers/login-handlers";
 import axios from "axios";
-import { populateDatabase } from "../src/seed";
 
 const getUserByID = `#graphql
   query GetUser ($id: Int!) {
@@ -16,40 +15,13 @@ const getUserByID = `#graphql
     }
   }`;
 
-async function getUserByIDRequest(url: string, id: number, token: string) {
+async function serverRequest(url: string, id: number, token: string) {
   const response = await axios.post(
     url,
     {
       query: getUserByID,
       variables: {
         id
-      }
-    },
-    {
-      headers: {
-        Authorization: token
-      }
-    }
-  );
-  return response.data;
-}
-
-const getManyUsers = `#graphql 
-  query GetManyUsers ($limit: Int) {
-    getManyUsers(limit: $limit) {
-      id
-      name
-      email
-    }
-  }`;
-
-async function getManyUsersRequest(url: string, limit: number, token: string) {
-  const response = await axios.post(
-    url,
-    {
-      query: getManyUsers,
-      variables: {
-        limit
       }
     },
     {
@@ -75,7 +47,7 @@ describe("User Query Test", function () {
     });
 
     const token = generateToken(userDB.id, false);
-    const userQueryResponse = await getUserByIDRequest(url, userDB.id, token);
+    const userQueryResponse = await serverRequest(url, userDB.id, token);
 
     const expectedUserDB = userDB && {
       id: userDB.id,
@@ -100,7 +72,7 @@ describe("User Query Test", function () {
     });
 
     const token = generateToken(userDB.id, false);
-    const userQueryResponse = await getUserByIDRequest(url, 1, token);
+    const userQueryResponse = await serverRequest(url, 1, token);
 
     expect(userQueryResponse.errors).to.be.deep.eq([
       {
@@ -123,38 +95,8 @@ describe("User Query Test", function () {
     });
 
     const token = generateToken(userDB.id, false);
-    const userQueryResponse = await getUserByIDRequest(url, userDB.id, "invalidToken123");
+    const userQueryResponse = await serverRequest(url, userDB.id, "invalidToken123");
 
-    expect(userQueryResponse.errors).to.be.deep.eq([
-      {
-        code: 401,
-        message: "Operação não autorizada"
-      }
-    ]);
-  });
-
-  it("Tried to find many users in the database", async function () {
-    await populateDatabase(7);
-
-    const token = generateToken(1, false);
-
-    const userQueryResponse = await getManyUsersRequest(url, 5, token);
-    expect(userQueryResponse.data.getManyUsers.length).to.be.equal(5);
-  });
-
-  it("Tried to find more users than there are in the database", async function () {
-    await populateDatabase(3);
-
-    const token = generateToken(1, false);
-
-    const userQueryResponse = await getManyUsersRequest(url, 5, token);
-    expect(userQueryResponse.data.getManyUsers.length).to.be.equal(3);
-  });
-
-  it("Tried to find user with an invalid token", async function () {
-    await populateDatabase(7);
-
-    const userQueryResponse = await getManyUsersRequest(url, 5, "invalidtoken123");
     expect(userQueryResponse.errors).to.be.deep.eq([
       {
         code: 401,
